+++ compiler/src/org.graalvm.compiler.truffle.common/src/org/graalvm/compiler/truffle/common/CompilableTruffleAST.java
@@ -60,8 +60,9 @@ public interface CompilableTruffleAST {
      * @param permanentBailout specifies if a bailout is due to a condition that probably won't
      *            change if this AST is compiled again. This value is meaningless if
      *            {@code bailout == false}.
+     * @param graphTooBig graph was too big
      */
-    void onCompilationFailed(Supplier<String> serializedException, boolean bailout, boolean permanentBailout);
+    void onCompilationFailed(Supplier<String> serializedException, boolean bailout, boolean permanentBailout, boolean graphTooBig);
 
     /**
      * Gets a descriptive name for this call target.