+++ truffle/src/com.oracle.truffle.api/src/com/oracle/truffle/api/impl/Accessor.java
@@ -56,7 +56,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.locks.Lock;
-import java.util.function.BiFunction;
 import java.util.function.Function;
 import java.util.function.Supplier;
 import java.util.logging.Handler;
@@ -100,6 +99,7 @@ import com.oracle.truffle.api.nodes.LanguageInfo;
 import com.oracle.truffle.api.nodes.Node;
 import com.oracle.truffle.api.nodes.NodeInterface;
 import com.oracle.truffle.api.nodes.RootNode;
+import com.oracle.truffle.api.nodes.ExecutionSignature;
 import com.oracle.truffle.api.source.Source;
 import com.oracle.truffle.api.source.Source.SourceBuilder;
 import com.oracle.truffle.api.source.SourceSection;
@@ -172,6 +172,8 @@ public abstract class Accessor {
         public abstract boolean isTrivial(RootNode rootNode);
 
         public abstract Object translateStackTraceElement(TruffleStackTraceElement stackTraceLement);
+
+        public abstract ExecutionSignature prepareForAOT(RootNode rootNode);
     }
 
     public abstract static class SourceSupport extends Support {
@@ -186,9 +188,8 @@ public abstract class Accessor {
 
         public abstract Source copySource(Source source);
 
-        public abstract void setPolyglotSource(Source source, org.graalvm.polyglot.Source polyglotSource);
-
-        public abstract org.graalvm.polyglot.Source getPolyglotSource(Source source);
+        public abstract org.graalvm.polyglot.Source getOrCreatePolyglotSource(Source source,
+                        Function<Source, org.graalvm.polyglot.Source> createSource);
 
         public abstract String findMimeType(URL url, Object fileSystemContext) throws IOException;
 
@@ -197,6 +198,8 @@ public abstract class Accessor {
         public abstract void setFileSystemContext(SourceBuilder builder, Object fileSystemContext);
 
         public abstract void invalidateAfterPreinitialiation(Source source);
+
+        public abstract void mergeLoadedSources(Source[] sources);
     }
 
     public abstract static class InteropSupport extends Support {
@@ -268,6 +271,8 @@ public abstract class Accessor {
         public abstract Env getEnvForInstrument(LanguageInfo language);
 
         public abstract Env getLegacyLanguageEnv(Object obj, boolean nullForHost);
+
+        public abstract boolean hasCurrentContext();
 
         public abstract ContextReference<Object> getCurrentContextReference(Object polyglotLanguage);
 
@@ -355,7 +360,7 @@ public abstract class Accessor {
 
         public abstract PolyglotException wrapGuestException(String languageId, Throwable exception);
 
-        public abstract <T> T getOrCreateRuntimeData(Object polyglotEngine, BiFunction<OptionValues, Function<String, TruffleLogger>, T> constructor);
+        public abstract <T> T getOrCreateRuntimeData(Object polyglotEngine);
 
         public abstract Set<? extends Class<?>> getProvidedTags(LanguageInfo language);
 
@@ -514,6 +519,13 @@ public abstract class Accessor {
 
         public abstract Object legacyScopes2ScopeObject(NodeInterface node, Iterator<com.oracle.truffle.api.Scope> legacyScopes, Class<? extends TruffleLanguage<?>> language);
 
+        public abstract OptionValues getEngineOptionValues(Object polyglotEngine);
+
+        public abstract Collection<? extends CallTarget> findCallTargets(Object polyglotEngine);
+
+        public abstract void preinitializeContext(Object polyglotEngine);
+
+        public abstract Object getEngineLock(Object polyglotEngine);
     }
 
     public abstract static class LanguageSupport extends Support {
@@ -676,11 +688,12 @@ public abstract class Accessor {
 
         public abstract <T> T getInstrumentationHandlerService(Object handler, Object polyglotInstrument, Class<T> type);
 
-        public abstract Object createInstrumentationHandler(Object polyglotEngine, DispatchOutputStream out, DispatchOutputStream err, InputStream in, MessageTransport messageInterceptor);
+        public abstract Object createInstrumentationHandler(Object polyglotEngine, DispatchOutputStream out, DispatchOutputStream err, InputStream in, MessageTransport messageInterceptor,
+                        boolean strongReferences);
 
         public abstract void collectEnvServices(Set<Object> collectTo, Object polyglotLanguage, TruffleLanguage<?> language);
 
-        public abstract void onFirstExecution(RootNode rootNode);
+        public abstract void onFirstExecution(RootNode rootNode, boolean validate);
 
         public abstract void onLoad(RootNode rootNode);
 
@@ -745,6 +758,8 @@ public abstract class Accessor {
         public abstract void notifyEnter(Object instrumentationHandler, TruffleContext truffleContext);
 
         public abstract void notifyLeave(Object instrumentationHandler, TruffleContext truffleContext);
+
+        public abstract Collection<CallTarget> getLoadedCallTargets(Object instrumentationHandler);
 
     }
 
@@ -836,7 +851,7 @@ public abstract class Accessor {
         /**
          * Returns the compiler options specified available from the runtime.
          */
-        public abstract OptionDescriptors getCompilerOptionDescriptors();
+        public abstract OptionDescriptors getEngineOptionDescriptors();
 
         /**
          * Returns <code>true</code> if the java stack frame is a representing a guest language
@@ -848,10 +863,6 @@ public abstract class Accessor {
         public abstract void initializeProfile(CallTarget target, Class<?>[] argumentTypes);
 
         public abstract <T extends Node> BlockNode<T> createBlockNode(T[] elements, ElementExecutor<T> executor);
-
-        public abstract void reloadEngineOptions(Object runtimeData, OptionValues optionValues);
-
-        public abstract void onEngineClosed(Object runtimeData);
 
         public abstract String getSavedProperty(String key);
 
@@ -869,6 +880,20 @@ public abstract class Accessor {
         public abstract boolean inFirstTier();
 
         public abstract void flushCompileQueue(Object runtimeData);
+
+        public abstract Object createRuntimeData(OptionValues options, Function<String, TruffleLogger> loggerFactory);
+
+        public abstract Object tryLoadCachedEngine(OptionValues runtimeData, Function<String, TruffleLogger> logger);
+
+        public abstract void onEngineCreate(Object engine, Object runtimeData);
+
+        public abstract boolean isStoreEnabled(OptionValues options);
+
+        public abstract void onEnginePatch(Object runtimeData, OptionValues options, Function<String, TruffleLogger> logSupplier);
+
+        public abstract boolean onEngineClosing(Object runtimeData);
+
+        public abstract void onEngineClosed(Object runtimeData);
 
     }
 
@@ -964,8 +989,9 @@ public abstract class Accessor {
             case "com.oracle.truffle.api.io.IOAccessor":
             case "com.oracle.truffle.api.frame.FrameAccessor":
             case "com.oracle.truffle.polyglot.EngineAccessor":
-            case "com.oracle.truffle.api.utilities.JSONHelper.DumpAccessor":
-                // OK, classes initializing accessors
+            case "com.oracle.truffle.api.utilities.JSONHelper.DumpAccessor":// OK, classes
+                                                                            // initializing
+                                                                            // accessors
                 break;
             case "com.oracle.truffle.api.debug.Debugger$AccessorDebug":
             case "com.oracle.truffle.tck.instrumentation.VerifierInstrument$TruffleTCKAccessor":
@@ -975,8 +1001,9 @@ public abstract class Accessor {
             case "com.oracle.truffle.api.impl.DefaultRuntimeAccessor":
             case "org.graalvm.compiler.truffle.runtime.GraalRuntimeAccessor":
             case "org.graalvm.compiler.truffle.runtime.debug.CompilerDebugAccessor":
-            case "com.oracle.truffle.api.library.LibraryAccessor":
-                // OK, classes allowed to use accessors
+            case "com.oracle.svm.enterprise.truffle.SubstrateEnterpriseAccessor":
+            case "com.oracle.truffle.api.library.LibraryAccessor":// OK, classes allowed to use
+                                                                  // accessors
                 break;
             default:
                 throw new IllegalStateException(this.getClass().getName());